feat: migrate ResearcherAccordionRow.spec
